backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic[email]==2.9.0
pydantic-settings==2.6.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.13.3

# Task Queue
# ============================================================
# 🔴 CELERY + REDIS — PRESERVED, NOT INSTALLED BY DEFAULT
# Uncomment these when reactivating the Celery approach.
# ============================================================
# celery==5.4.0
# redis==5.2.0

# Scheduler
# ✅ ACTIVE — Scheduler (replaces Celery Beat)
apscheduler==3.10.4

# HTTP & Scraping
httpx==0.27.2
beautifulsoup4==4.12.3
playwright==1.48.0
python-whois==0.9.4
dnspython==2.7.0

# AI
groq==0.11.0
google-generativeai>=0.8.3

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# PDF
reportlab==4.2.5
weasyprint==62.3

# Excel / Reports
openpyxl==3.1.5
matplotlib==3.9.2

# Notifications
python-telegram-bot==21.6

# Utilities
loguru==0.7.2
python-dotenv==1.0.1
slowapi==0.1.9
tenacity==9.0.0       # Retry logic
python-multipart==0.0.12
bleach==6.1.0         # XSS Sanitization

# Authentication
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Supabase
supabase==2.10.0

# Payment Gateway
razorpay==1.4.2
